--- packages/forgae-cli/forgae-init/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "forgae-init",
-  "version": "1.0.0",
-  "description": "",
-  "main": "init.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "forgae-utils": "1.0.0"
-  }
-}
