+++ packages/forgae-lib/forgae-deployer.js
@@ -1,7 +1,7 @@
-const utils = require('forgae-utils/utils')
+const utils = require('forgae-utils')
 const fs = require('fs');
-const logStoreService = require('./../forgae-history/log-store-service');
-const config = require('./../config.json');
+const logStoreService = require('forgae-logger').history;
+const config = require('forgae-config');
 const decodedHexAddressToPublicAddress = utils.decodedHexAddressToPublicAddress;
 let ttl = 100;
 const opts = {
