+++ packages/forgae-logger/index.js
@@ -1,0 +1,5 @@
+const history = require('./logger-service/log-store-service');
+
+module.exports = {
+    history
+}
