+++ packages/forgae-logger/package.json
@@ -2,7 +2,7 @@
   "name": "forgae-logger",
   "version": "1.0.0",
   "description": "",
-  "main": "utils.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
