--- packages/forgae-utils/utils/test-utils.js
@@ -1,46 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-
-const readFile = async (path, encoding = null, errTitle = 'READ FILE ERR') => {
-    try {
-        return fs.readFileSync(
-            path,
-            encoding
-        )
-    } catch (e) {
-        switch (e.code) {
-            case 'ENOENT':
-                throw new Error('File not found')
-            default:
-                throw e
-        }
-    }
-};
-
-const writeFile = (path, content) => {
-	fs.writeFileSync(path, content);
-}
-
-const writeFileRelative = async (relativePath, content = null) => {
-	return writeFile(path.resolve(process.cwd(), relativePath), content);
-}
-
-const readFileRelative = (relativePath, encoding = null, errTitle = 'READ FILE ERR') => {
-	return readFile(path.resolve(process.cwd(), relativePath), encoding, errTitle);
-}
-
-const fileExists = (relativePath) => {
-	return fs.existsSync(path.resolve(process.cwd(), relativePath));
-}
-
-const trimAdresseses = (addressToTrim) => {
-	return addressToTrim.substring(3)
-}
-
-module.exports = {
-	readFile,
-	readFileRelative,
-	writeFileRelative,
-	fileExists,
-	trimAdresseses
-}
